refactor(version): decompose version.rs into submodule directory
Split the monolithic src/version.rs (3,051 lines) into a well-organized
src/version/ directory with 15 focused submodule files:

  version/
  ├── mod.rs              # Module declarations + re-exports
  ├── types.rs            # VersionInfo, ConflictEntry, InstallResult, ChecksumResult
  ├── cache.rs            # Embedded version list load/save + merge logic
  ├── compare.rs          # version_compare, version_less_than
  ├── manager.rs          # VersionManager struct + core impl (NPM, checksums, streaming)
  ├── claude.rs           # Claude Code GCS install, conflict detection, checksums
  ├── codex.rs            # Codex GitHub releases, binary install
  ├── gemini.rs           # Gemini NPM install
  ├── pi.rs               # Pi NPM install
  ├── hermes.rs           # Hermes git clone + bash installer
  ├── opencode.rs         # OpenCode upgrade + NPM fallback
  ├── antigravity.rs      # AUR helper install
  ├── unleash_self.rs     # Self-update version list
  ├── cli.rs              # list_versions, install_version, show_current CLI entries
  └── tests.rs            # All 12 unit tests (passing)

All existing public API paths (crate::version::*) are preserved via
re-exports in mod.rs. No logic changes — purely structural.

Follows the same pattern as the agents.rs decomposition (#428).
